refactor(observability): 分离体验报告派生职责
